Rename and move SpringSecurityCoreVersion2 to Version
