+++ oauth2-authorization-server/src/main/java/org/springframework/security/oauth2/server/authorization/config/Settings.java
@@ -15,7 +15,7 @@
  */
 package org.springframework.security.oauth2.server.authorization.config;
 
-import org.springframework.security.core.SpringSecurityCoreVersion2;
+import org.springframework.security.oauth2.server.authorization.Version;
 import org.springframework.util.Assert;
 
 import java.io.Serializable;
@@ -30,7 +30,7 @@ import java.util.function.Consumer;
  * @since 0.0.2
  */
 public class Settings implements Serializable {
-	private static final long serialVersionUID = SpringSecurityCoreVersion2.SERIAL_VERSION_UID;
+	private static final long serialVersionUID = Version.SERIAL_VERSION_UID;
 	private final Map<String, Object> settings;
 
 	/**
